feat(CustomGroups): code cleanup and added logic for custom groups
