+++ src/prerna/auth/AccessToken.java
@@ -1,6 +1,7 @@
 package prerna.auth;
 
 import java.io.Serializable;
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -16,7 +17,6 @@ public class AccessToken implements Serializable {
 	// this will store all the groups that the user has
 	// will be provided to us when the user logs in 
 	// from an IDP
-	Collection<String> userCustomGroups = null;
 	Collection<String> userGroups = null;
 	String userGroupType = null;
 	
@@ -42,6 +42,7 @@ public class AccessToken implements Serializable {
 	String modelUsageRestriction = null;
 	
 	Map<String, String> sans = null;
+	
 	Map<String, Collection<String>> meta = null;
 	
 	boolean locked = false;
@@ -87,14 +88,6 @@ public class AccessToken implements Serializable {
 
 	public void setUserGroupType(String userGroupType) {
 		this.userGroupType = userGroupType;
-	}
-
-	public Collection<String> getUserCustomGroups() {
-		return userCustomGroups;
-	}
-
-	public void setUserCustomGroups(Set<String> userCustomGroups) {
-		this.userCustomGroups = userCustomGroups;
 	}
 
 	public void setExpires_in(int expires_in) {
@@ -279,4 +272,5 @@ public class AccessToken implements Serializable {
 		this.modelUsageRestriction = modelUsageRestriction;
 	}
 
+	
 }
