+++ src/prerna/auth/utils/SecurityGroupEngineUtils.java
@@ -56,12 +56,10 @@ public class SecurityGroupEngineUtils extends AbstractSecurityUtils {
 				// one of the logins has a group. set checker to false
 				anyUserGroups=true;
 			}
-			if (user.getAccessToken(login).getUserCustomGroups() != null && !user.getAccessToken(login).getUserCustomGroups().isEmpty()) {
-		        AndQueryFilter customAndFilter = new AndQueryFilter();
-		        customAndFilter.addFilter(SimpleQueryFilter.makeColToValFilter("GROUPENGINEPERMISSION__TYPE", "==", "CUSTOM"));
-		        customAndFilter.addFilter(SimpleQueryFilter.makeColToValFilter("GROUPENGINEPERMISSION__ID", "==", user.getAccessToken(login).getUserCustomGroups()));
-		        orFilter.addFilter(customAndFilter);
-		    }
+			AndQueryFilter customAndFilter = new AndQueryFilter();
+			customAndFilter.addFilter(SimpleQueryFilter.makeColToValFilter("CUSTOMGROUPASSIGNMENT__USERID", "==", user.getAccessToken(login).getId()));
+			orFilter.addFilter(customAndFilter);
+			
 			AndQueryFilter andFilter = new AndQueryFilter();
 			andFilter.addFilter(SimpleQueryFilter.makeColToValFilter("GROUPENGINEPERMISSION__TYPE", "==", user.getAccessToken(login).getUserGroupType()));
 			andFilter.addFilter(SimpleQueryFilter.makeColToValFilter("GROUPENGINEPERMISSION__ID", "==", user.getAccessToken(login).getUserGroups()));
@@ -133,12 +131,10 @@ public class SecurityGroupEngineUtils extends AbstractSecurityUtils {
 				// one of the logins has a group. set checker to false
 				anyUserGroups=true;
 			}
-			if (user.getAccessToken(login).getUserCustomGroups() != null && !user.getAccessToken(login).getUserCustomGroups().isEmpty()) {
-		        AndQueryFilter customAndFilter = new AndQueryFilter();
-		        customAndFilter.addFilter(SimpleQueryFilter.makeColToValFilter("GROUPENGINEPERMISSION__TYPE", "==", "CUSTOM"));
-		        customAndFilter.addFilter(SimpleQueryFilter.makeColToValFilter("GROUPENGINEPERMISSION__ID", "==", user.getAccessToken(login).getUserCustomGroups()));
-		        orFilter.addFilter(customAndFilter);
-		    }
+			AndQueryFilter customAndFilter = new AndQueryFilter();
+			customAndFilter.addFilter(SimpleQueryFilter.makeColToValFilter("CUSTOMGROUPASSIGNMENT__USERID", "==", user.getAccessToken(login).getId()));
+			orFilter.addFilter(customAndFilter);
+			
 			AndQueryFilter andFilter = new AndQueryFilter();
 			andFilter.addFilter(SimpleQueryFilter.makeColToValFilter("GROUPENGINEPERMISSION__TYPE", "==", user.getAccessToken(login).getUserGroupType()));
 			andFilter.addFilter(SimpleQueryFilter.makeColToValFilter("GROUPENGINEPERMISSION__ID", "==", user.getAccessToken(login).getUserGroups()));
@@ -210,12 +206,10 @@ public class SecurityGroupEngineUtils extends AbstractSecurityUtils {
 				// one of the logins has a group. set checker to false
 				anyUserGroups=true;
 			}
-			if (user.getAccessToken(login).getUserCustomGroups() != null && !user.getAccessToken(login).getUserCustomGroups().isEmpty()) {
-		        AndQueryFilter customAndFilter = new AndQueryFilter();
-		        customAndFilter.addFilter(SimpleQueryFilter.makeColToValFilter("GROUPENGINEPERMISSION__TYPE", "==", "CUSTOM"));
-		        customAndFilter.addFilter(SimpleQueryFilter.makeColToValFilter("GROUPENGINEPERMISSION__ID", "==", user.getAccessToken(login).getUserCustomGroups()));
-		        orFilter.addFilter(customAndFilter);
-		    }
+			AndQueryFilter customAndFilter = new AndQueryFilter();
+			customAndFilter.addFilter(SimpleQueryFilter.makeColToValFilter("CUSTOMGROUPASSIGNMENT__USERID", "==", user.getAccessToken(login).getId()));
+			orFilter.addFilter(customAndFilter);
+			
 			AndQueryFilter andFilter = new AndQueryFilter();
 			andFilter.addFilter(SimpleQueryFilter.makeColToValFilter("GROUPENGINEPERMISSION__TYPE", "==", user.getAccessToken(login).getUserGroupType()));
 			andFilter.addFilter(SimpleQueryFilter.makeColToValFilter("GROUPENGINEPERMISSION__ID", "==", user.getAccessToken(login).getUserGroups()));
